requirements.txt
Flask
Flask-Cors
Pillow
stegano
pycryptodome
piexif
gunicorn
python-dotenv